requirements.txt

streamlit
pandas
sentence-transformers
pdfminer.six
PyMuPDF
plotly
reportlab
